createBroccoliInitOptions(), createBroccoli() を追加
- 外部から、broccoliに渡す設定情報だけを取り出したい場合などを想定して機能分解した。
