+++ src/editor/broccoli/broccoli.js
@@ -32,8 +32,6 @@ module.exports = function(px2ce){
 	this.init = function(editorOption, callback){
 		callback = callback || function(){};
 
-		var customFields = {};
-
 		new Promise(function(rlv){rlv();})
 			.then(function(){ return new Promise(function(rlv, rjt){
 				px2ce.gpiBridge(
@@ -42,35 +40,36 @@ module.exports = function(px2ce){
 					},
 					function(_px2conf){
 						px2conf = _px2conf;
-
-						toolbar.init({
-							"btns":[
-								{
-									"label": px2ce.lb.get('ui_label.toggle_instance_treeview'),
-									"click": function(){
-										show_instanceTreeView = (show_instanceTreeView ? false : true);
-										_this.redraw(function(){
-											// alert('完了');
-										});
-									}
-								},
-								{
-									"label": px2ce.lb.get('ui_label.open_in_browser'),
-									"click": function(){
-										px2ce.openUrlInBrowser( getPreviewUrl() );
-									}
-								}
-							],
-							"onFinish": function(){
-								// 完了イベント
-								px2ce.finish();
-							}
-						},function(){
-							rlv();
-						});
-
+						rlv();
 					}
 				);
+			}); })
+			.then(function(){ return new Promise(function(rlv, rjt){
+				toolbar.init({
+					"btns":[
+						{
+							"label": px2ce.lb.get('ui_label.toggle_instance_treeview'),
+							"click": function(){
+								show_instanceTreeView = (show_instanceTreeView ? false : true);
+								_this.redraw(function(){
+									// alert('完了');
+								});
+							}
+						},
+						{
+							"label": px2ce.lb.get('ui_label.open_in_browser'),
+							"click": function(){
+								px2ce.openUrlInBrowser( getPreviewUrl() );
+							}
+						}
+					],
+					"onFinish": function(){
+						// 完了イベント
+						px2ce.finish();
+					}
+				},function(){
+					rlv();
+				});
 			}); })
 			.then(function(){ return new Promise(function(rlv, rjt){
 				$canvas.append((function(){
@@ -100,104 +99,20 @@ module.exports = function(px2ce){
 				rlv();
 			}); })
 			.then(function(){ return new Promise(function(rlv, rjt){
-				// プロジェクトが拡張するフィールド
-				// クライアントサイドのライブラリをロードしておく
-				px2ce.gpiBridge(
-					{
-						'api': 'loadCustomFieldsClientSideLibs'
-					},
-					function(scripts){
-						for(var i in scripts){
-							$('body').append(scripts[i]);
-						}
-						rlv();
-					}
-				);
-
-			}); })
-			.then(function(){ return new Promise(function(rlv, rjt){
-				// フィールドを拡張
-
-				// px2ce が拡張するフィールド
-				customFields.table = window.BroccoliFieldTable;
-
-				// 呼び出し元アプリが拡張するフィールド
-				for( var idx in px2ce.options.customFields ){
-					customFields[idx] = px2ce.options.customFields[idx];
-				}
-
-				// プロジェクトが拡張するフィールド
-				var confCustomFields = {};
-				try {
-					confCustomFields = px2conf.plugins.px2dt.guieditor.custom_fields;
-					for( var fieldName in confCustomFields ){
-						try {
-							if( confCustomFields[fieldName].frontend.file && confCustomFields[fieldName].frontend.function ){
-								// console.log(eval( confCustomFields[fieldName].frontend.function ));
-								customFields[fieldName] = eval( confCustomFields[fieldName].frontend.function );
-							}else{
-								console.error( 'FAILED to load custom field: ' + fieldName + ' (frontend);' );
-								console.error( 'unknown type' );
-							}
-						} catch (e) {
-							console.error( 'FAILED to load custom field: ' + fieldName + ' (frontend);' );
-							console.error(e);
-						}
-					}
-				} catch (e) {
-				}
-
-				// console.log(customFields);
-
-				rlv();
-			}); })
-			.then(function(){ return new Promise(function(rlv, rjt){
-
+				// broccoli-html-editor オブジェクトを生成
 				broccoli = new Broccoli();
-				broccoli.init(
-					{
-						'elmCanvas': $elmCanvas.get(0),
-						'elmModulePalette': $elmModulePalette.get(0),
-						'elmInstanceTreeView': $elmInstanceTreeView.get(0),
-						'elmInstancePathView': $elmInstancePathView.get(0),
-						'contents_area_selector': px2conf.plugins.px2dt.contents_area_selector,
-						// ↑編集可能領域を探すためのクエリを設定します。
-						//  この例では、data-contents属性が付いている要素が編集可能領域として認識されます。
-						'contents_bowl_name_by': px2conf.plugins.px2dt.contents_bowl_name_by,
-						// ↑bowlの名称を、data-contents属性値から取得します。
-						'customFields': customFields,
-						'lang': px2ce.options.lang,
-						'gpiBridge': function(api, options, callback){
-							// GPI(General Purpose Interface) Bridge
-							// broccoliは、バックグラウンドで様々なデータ通信を行います。
-							// GPIは、これらのデータ通信を行うための汎用的なAPIです。
-							px2ce.gpiBridge(
-								{
-									'api': 'broccoliBridge',
-									'page_path': page_path,
-									'forBroccoli':{
-										'api': api,
-										'options': options
-									}
-								},
-								function(data){
-									callback(data);
-								}
-							);
-							return;
-						},
-						'onClickContentsLink': function( uri, data ){
-							px2ce.onClickContentsLink( uri, data );
-						},
-						'onMessage': function( message ){
-							// ユーザーへ知らせるメッセージを表示する
-							px2ce.message(message);
+				px2ce.createBroccoliInitOptions(function( broccoliInitOptions ){
+					broccoliInitOptions.elmCanvas = $elmCanvas.get(0),
+					broccoliInitOptions.elmModulePalette = $elmModulePalette.get(0),
+					broccoliInitOptions.elmInstanceTreeView = $elmInstanceTreeView.get(0),
+					broccoliInitOptions.elmInstancePathView = $elmInstancePathView.get(0),
+					broccoli.init(
+						broccoliInitOptions ,
+						function(){
+							rlv();
 						}
-					} ,
-					function(){
-						rlv();
-					}
-				);
+					);
+				});
 			}); })
 			.then(function(){ return new Promise(function(rlv, rjt){
 				// 初期化が完了すると呼びだされるコールバック関数です。
